pyproject.toml
[project]
name = "mdlens"
version = "0.1.0"
description = "A lightweight read-only Markdown reader with a SQLite search index."
readme = "README.md"
requires-python = ">=3.12"
keywords = ["fastapi", "markdown", "reader", "sqlite", "viewer"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Environment :: Web Environment",
    "Intended Audience :: End Users/Desktop",
    "Operating System :: Microsoft :: Windows",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3 :: Only",
    "Programming Language :: Python :: 3.12",
    "Topic :: Text Processing :: Markup :: Markdown",
]
dependencies = [
    "fastapi>=0.115,<1.0",
    "markdown-it-py>=3.0,<4.0",
    "pydantic>=2.0,<3.0",
    "sqlalchemy>=2.0,<3.0",
    "uvicorn>=0.30,<1.0",
]

[dependency-groups]
dev = [
    "httpx>=0.27,<1.0",
    "pytest>=8.0,<9.0",
    "pytest-cov>=5.0,<8.0",
]

[project.scripts]
mdlens = "mdlens:main"

[project.urls]
Homepage = "https://github.com/3pintea/MdLens"
Repository = "https://github.com/3pintea/MdLens"
Issues = "https://github.com/3pintea/MdLens/issues"

[[tool.uv.index]]
name = "testpypi"
url = "https://test.pypi.org/simple/"
publish-url = "https://test.pypi.org/legacy/"
explicit = true

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = [
    "--cov=mdlens",
    "--cov-report=term-missing",
    "--cov-fail-under=90",
    "-p",
    "no:cacheprovider",
]

[tool.coverage.run]
branch = true
source = ["mdlens"]
omit = [
    "src/mdlens/__main__.py",
    "src/mdlens/ui.py",
]

[tool.coverage.report]
show_missing = true
skip_covered = false

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/mdlens"]

[tool.hatch.build.targets.sdist]
include = [
    "/docs",
    "/src",
    "/tests",
    "/README.md",
    "/pyproject.toml",
]
exclude = [
    "/.pytest-tmp",
    "/pytest-cache-files-*",
    "/pytest_tmp_workspace",
    "/test_tmp_workspace",
]
